# bin/Release/net6.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "PruebaTropigasAPI, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
    "/api/Configuracion": {
      "get": {
        "tags": [
          "Ventas"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Articulos": {
      "get": {
        "tags": [
          "Ventas"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Venta": {
      "get": {
        "tags": [
          "Ventas"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "post": {
        "tags": [
          "Ventas"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Venta"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/VentaD": {
      "get": {
        "tags": [
          "Ventas"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "post": {
        "tags": [
          "Ventas"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/VentaD"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Venta/{id}": {
      "put": {
        "tags": [
          "Ventas"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Venta"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Ventas"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/VentaD/{id}": {
      "put": {
        "tags": [
          "Ventas"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/VentaD"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Ventas"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    }
  },
  "definitions": {
    "Venta": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "fecha": {
          "format": "date-time",
          "type": "string"
        },
        "importe": {
          "format": "double",
          "type": "number"
        },
        "usuario": {
          "type": "string"
        },
        "enviado": {
          "type": "boolean"
        }
      }
    },
    "VentaD": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "renglon": {
          "format": "int32",
          "type": "integer"
        },
        "articulo": {
          "type": "string"
        },
        "cantidad": {
          "format": "int32",
          "type": "integer"
        },
        "precio": {
          "format": "double",
          "type": "number"
        },
        "enviado": {
          "type": "boolean"
        }
      }
    }
  }
}